Converted Data types in CBSE Board Calculator some unnessary float to int

+++ src/com/company/Exercise_1_BoardPercentCalculator.java
@@ -6,30 +6,30 @@ public class Exercise_1_BoardPercentCalculator
 {
     public static void main(String[] args)
     {
-        float Sub_1;
-        float Sub_2;
-        float Sub_3;
-        float Sub_4;
-        float Sub_5;
-        float total_Marks;
+        int Sub_1;
+        int Sub_2;
+        int Sub_3;
+        int Sub_4;
+        int Sub_5;
+        int total_Marks;
         float percentage;
 
         Scanner take = new Scanner(System.in);
 
         System.out.print("Enter Marks for Subject 1: ");
-        Sub_1 = take.nextFloat();
+        Sub_1 = take.nextInt();
 
         System.out.print("Enter Marks for Subject 2: ");
-        Sub_2 = take.nextFloat();
+        Sub_2 = take.nextInt();
 
         System.out.print("Enter Marks for Subject 3: ");
-        Sub_3 = take.nextFloat();
+        Sub_3 = take.nextInt();
 
         System.out.print("Enter Marks for Subject 4: ");
-        Sub_4 = take.nextFloat();
+        Sub_4 = take.nextInt();
 
         System.out.print("Enter Marks for Subject 5: ");
-        Sub_5 = take.nextFloat();
+        Sub_5 = take.nextInt();
 
         total_Marks = Sub_1 + Sub_2 + Sub_3 + Sub_4 + Sub_5;
         percentage = (total_Marks * 100) / 500.0F;
